Support for Swift 4.2 (#268)
* Update to Swift 4.2 (#265)

* Update to Swift 4.2

Swift 4.2 moves to split out some of the ApplicationDelegate keys to be named better, this implements that. It solves build errors when using the Xcode 10 GM build.

Fixes #262

* Set Travis to run Xcode 10

* Set Swift version to 4.2

By doing this the library no longer shows as “Convert to Swift 4.2” in Xcode

* Update to Swift 4.2 (#265)

* Update to Swift 4.2

Swift 4.2 moves to split out some of the ApplicationDelegate keys to be named better, this implements that. It solves build errors when using the Xcode 10 GM build.

Fixes #262

* Set Travis to run Xcode 10

* Set Swift version to 4.2

By doing this the library no longer shows as “Convert to Swift 4.2” in Xcode

* Upgraded workspace to Swift 4.2

* Updated Carthage deps

* Deprecated Swift components that wrapped deprecated FBSDK components